refactor: move toolbars to designated components
