+++ src/Toolbars/PaintLayerBar/index.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from "react";
 import { Button, Switch, Slider } from "@cognite/cogs.js";
-import CogniteFileViewerContext from "../Cognite/FileViewerContext";
+import CogniteFileViewerContext from "../../Cognite/FileViewerContext";
+import { DEFAULT } from "../../utils";
 import ColorPicker from "./ColorPicker";
 import { WrappingBar, BarSection } from "./components";
-import { DEFAULT } from "../utils";
 
-export const PaintLayerBar = (): JSX.Element => {
+export default function PaintLayerBar(): JSX.Element {
   const {
     paintLayerCanvasRef,
     paintLayerEditMode,
@@ -95,4 +95,4 @@ export const PaintLayerBar = (): JSX.Element => {
       </BarSection>
     </WrappingBar>
   );
-};
+}
